Adds Card class
Adds a class to represent a card.

As well as holding the standard card attibutes faceValue, suit and value
Cards are also comparible to other cards.

There's an equals() method which defines equality

It also implements the Comparable interface which forces us to implement compareTo
compareTo() defines how a Card should be compared to another Card.

Using these methods Cards can be sorted in collections or directly
compared to other Cards using isGreaterThan or is isLowerThan.

The not loads of code here but there's a few concepts going on for you
to have a play with and learn about.