Adds a super simple game loop
Adds a simple game loop which will play a round of BlackJack

A lot of features are missing for now. I just wanted to see
how I could use the classes I've made so far to run a game.

This code will all move out of Main and get much more polished.