Corrects Blackjack logic
The blackjack commit didn't work. I should write some tests
for Game but that seems hard, so I'll do it later.

What I did instead was add a method to the deck generator which
creates a deck of just A & 10. Which means even after a shuffle I didn't
have long to wait until I got one.